Refactor code to add functionality of properties file to get finite amount of coins
This change is made to add the code for getChangeFor: that
returns the optimal change for the passed amount and check
the coins from properties file which limits the number of coins
per denomination. This also contains the test cases to test the
various scenarios of the said method. There is still functionality
missing to save the revised amount of coins back in the properties
file. Also to test the error scenario in case there are no coins returned